feat: add a delete button

+++ back-end/src/repository/user.repository.ts
@@ -1,6 +1,7 @@
 
 
 import { PrismaClient } from '@prisma/client'
+import * as console from "console";
 
 const prisma = new PrismaClient()
 
@@ -21,6 +22,24 @@ async function fetchUser() {
 }
 
 async function deleteUser(user: string) {
+    console.log("deleteUser :", user);
+
+    const userToDelete = await prisma.user.findUnique({
+        where: {
+            name: user
+        }
+    })
+    if (!userToDelete) {
+        throw new Error(`User with name ${user} not found`)
+    }
+
+    // I didnt find a way to cascade delete the late entries so I delete them manually
+    await prisma.late.deleteMany({
+        where: {
+            userId: userToDelete.id
+        }
+    })
+
     await prisma.user.delete({
         where: {
             name: user
